picture_training_enhanced/requirements_training.txt
# Requirements pour l'entraînement de modèles de classification d'images
# Compatible avec GPU NVIDIA L4 et CUDA 11.8+

# PyTorch (installé séparément avec CUDA support)
# torch>=2.0.0
# torchvision>=0.15.0
# torchaudio>=2.0.0

# Core ML libraries
numpy>=1.21.0
scipy>=1.7.0
scikit-learn>=1.0.0
pandas>=1.3.0

# Vision et traitement d'images
Pillow>=9.0.0
opencv-python>=4.5.0
albumentations>=1.3.0

# Métriques et visualisation
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0
tensorboard>=2.10.0

# Deep Learning utilities
timm>=0.9.0  # PyTorch Image Models
efficientnet-pytorch>=0.7.1

# Progress bars et logging
tqdm>=4.62.0
colorlog>=6.6.0

# Configuration
pyyaml>=6.0
python-dotenv>=0.19.0

# Export de modèles (optionnel)
onnx>=1.12.0
onnxruntime>=1.12.0
# coremltools>=6.0  # Pour export iOS (installer si nécessaire)
# tensorflow>=2.10.0  # Pour export TFLite (installer si nécessaire)
# onnx-tf>=1.10.0  # Pour conversion ONNX->TF (installer si nécessaire)

# Optimisations
psutil>=5.9.0
gputil>=1.4.0

# Tests
pytest>=7.0.0
pytest-cov>=3.0.0